requirements.txt
numpy
pandas
scipy
scikit-learn
fastapi
uvicorn
python-multipart
openpyxl
statsmodels
python-jose
passlib
bcrypt==4.0.1
